This commit will add Data Context and Context Provider
