+++ ADMIN_DASHBOARD/src/Components/Tasks/TasksTable.jsx
@@ -28,7 +28,7 @@ export const TasksTable = ({ tasks, searchTerm, searchColumn,taskEdit }) => {
   });
 
   const filteredTasks = sortedTasks.filter((task) => {
-    console.log(tasks);
+    
     if (searchColumn === "Task Name")
       return task.title.toLowerCase().includes(searchTerm.toLowerCase());
     else if (searchColumn === "Company")
